+++ libraries/PixelGame/PixelRenderer.h
@@ -26,7 +26,7 @@ public:
     virtual void draw() = 0;
 
     virtual void setPixel(int x,int y,uint8_t r, uint8_t g , uint8_t b, uint8_t a) = 0;
-    virtual void setBrightness(int val) = 0;
+    virtual void setBrightness(uint8_t val) = 0;
     virtual void fade() = 0;
 };
